tweaks and cleanups in ZLIB.  Also, removed the use of Type.InvokeMember, which does not work on .NET CF.
git-svn-id: https://dotnetzip.svn.codeplex.com/svn@45646 4ef1bfeb-0a34-4d02-9305-240e5b07a168

+++ Zlib Tests/ZlibUnitTest1.cs
@@ -250,7 +250,7 @@ namespace Ionic.Zlib.Tests
             rc = compressor.SetDictionary(dictionary);
             Assert.AreEqual<int>(ZlibConstants.Z_OK, rc, String.Format("at SetDeflateDictionary() [{0}]", compressor.Message));
 
-            long dictId = compressor.Adler32;
+            int dictId = compressor.Adler32;
 
             compressor.OutputBuffer = compr;
             compressor.NextOut = 0;
